Load Launch Script Eager
